Split large module
